utils: split lyrics providers
