+++ crates/utils/src/lyrics.rs
@@ -2,18 +2,42 @@ use percent_encoding::NON_ALPHANUMERIC;
 use serde::Deserialize;
 use std::collections::{HashSet, hash_map::DefaultHasher};
 use std::hash::{Hash, Hasher};
-use std::sync::{Mutex, OnceLock};
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::time::{Duration, Instant};
 
 mod cache;
+mod local;
+mod lrc;
 mod model;
 mod request;
+mod server;
 
 use cache::{
     LyricsInflightGuard, load_persisted_lyrics, lyrics_cache, store_lyrics, try_begin_lyrics_fetch,
 };
+use local::fetch_local_lrc;
+use lrc::{extract_line_timestamps, parse_enhanced_words, parse_lrc};
 pub use model::{LyricChunk, LyricLine, Lyrics};
 pub use request::{LyricsRequest, LyricsServerAuth};
+use server::{fetch_jellyfin_lyrics, fetch_subsonic_lyrics};
+
+const LRCLIB_TIMEOUT: Duration = Duration::from_secs(5);
+const SERVER_LYRICS_TIMEOUT: Duration = Duration::from_secs(5);
+const LYRICS_INFLIGHT_POLL_INTERVAL: Duration = Duration::from_millis(50);
+const LYRICS_INFLIGHT_WAIT_TIMEOUT: Duration = Duration::from_secs(20);
+
+macro_rules! lyrics_debug {
+    ($($arg:tt)*) => {
+        if $crate::lyrics::lyrics_terminal_debug_enabled() {
+            tracing::debug!("[lyrics] {}", format_args!($($arg)*));
+        }
+    };
+}
+
+mod musixmatch;
+mod paxsenix;
+
+use musixmatch::fetch_from_musixmatch_enhanced;
+use paxsenix::{fetch_from_paxsenix_apple_music, fetch_from_paxsenix_youtube};
 
 #[derive(Debug, Deserialize)]
 struct LrcLibResponse {
@@ -21,108 +45,6 @@ struct LrcLibResponse {
     synced_lyrics: Option<String>,
     #[serde(rename = "plainLyrics")]
     plain_lyrics: Option<String>,
-}
-
-const MUSIXMATCH_ROOT_URL: &str = "https://apic-desktop.musixmatch.com/ws/1.1/";
-const PAXSENIX_ROOT_URL: &str = "https://lyrics.paxsenix.org";
-const ITUNES_SEARCH_ROOT_URL: &str = "https://itunes.apple.com/search";
-const MUSIXMATCH_TIMEOUT: Duration = Duration::from_secs(3);
-const PAXSENIX_TIMEOUT: Duration = Duration::from_secs(5);
-const PAXSENIX_APPLE_LYRICS_TIMEOUT: Duration = Duration::from_secs(10);
-const PAXSENIX_YOUTUBE_LYRICS_TIMEOUT: Duration = Duration::from_secs(3);
-const LRCLIB_TIMEOUT: Duration = Duration::from_secs(5);
-const SERVER_LYRICS_TIMEOUT: Duration = Duration::from_secs(5);
-const LYRICS_INFLIGHT_POLL_INTERVAL: Duration = Duration::from_millis(50);
-const LYRICS_INFLIGHT_WAIT_TIMEOUT: Duration = Duration::from_secs(20);
-
-static MUSIXMATCH_TOKEN: OnceLock<Mutex<Option<MusixmatchToken>>> = OnceLock::new();
-
-macro_rules! lyrics_debug {
-    ($($arg:tt)*) => {
-        if lyrics_terminal_debug_enabled() {
-            tracing::debug!("[lyrics] {}", format_args!($($arg)*));
-        }
-    };
-}
-
-struct MusixmatchToken {
-    value: String,
-    expires_at_ms: u128,
-}
-
-// --- Jellyfin lyrics types ---
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct JellyfinLyricsResponse {
-    #[serde(default)]
-    lyrics: Vec<JellyfinLyricLine>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct JellyfinLyricLine {
-    text: String,
-    start: Option<i64>,
-}
-
-// --- Subsonic lyrics types (getLyricsBySongId - OpenSubsonic extension) ---
-
-#[derive(Debug, Deserialize)]
-struct SubsonicSongLyricsEnvelope {
-    #[serde(rename = "subsonic-response")]
-    response: SubsonicSongLyricsResponse,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct SubsonicSongLyricsResponse {
-    status: String,
-    #[serde(default)]
-    lyrics_list: Option<SubsonicLyricsList>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct SubsonicLyricsList {
-    #[serde(default)]
-    structured_lyrics: Vec<SubsonicStructuredLyric>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct SubsonicStructuredLyric {
-    synced: bool,
-    #[serde(default)]
-    line: Vec<SubsonicLyricLine>,
-}
-
-#[derive(Debug, Deserialize)]
-struct SubsonicLyricLine {
-    #[serde(default)]
-    start: Option<i64>,
-    value: String,
-}
-
-// --- Subsonic lyrics types (getLyrics - basic endpoint) ---
-
-#[derive(Debug, Deserialize)]
-struct SubsonicPlainLyricsEnvelope {
-    #[serde(rename = "subsonic-response")]
-    response: SubsonicPlainLyricsResponse,
-}
-
-#[derive(Debug, Deserialize)]
-struct SubsonicPlainLyricsResponse {
-    status: String,
-    #[serde(default)]
-    lyrics: Option<SubsonicPlainLyricsData>,
-}
-
-#[derive(Debug, Deserialize)]
-struct SubsonicPlainLyricsData {
-    #[serde(default)]
-    value: String,
 }
 
 // --- Apple Music lyrics types ---
@@ -882,968 +804,6 @@ fn extract_server_id(path: &str, prefix: &str) -> Option<String> {
         .map(|s| s.to_string())
 }
 
-#[cfg(not(target_arch = "wasm32"))]
-async fn fetch_local_lrc(audio_path: &str) -> Option<Lyrics> {
-    let content = read_local_lrc(audio_path).or_else(|| read_embedded_lyrics(audio_path))?;
-    if content.trim().is_empty() {
-        return None;
-    }
-    let lines = if lrc_has_usable_timing(&content) {
-        parse_lrc(&content)
-    } else {
-        Vec::new()
-    };
-    if has_usable_line_timing(&lines) {
-        Some(Lyrics::Synced(lines))
-    } else {
-        Some(Lyrics::Plain(content))
-    }
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-fn read_embedded_lyrics(audio_path: &str) -> Option<String> {
-    use lofty::file::TaggedFileExt;
-    use lofty::probe::Probe;
-    use lofty::tag::ItemKey;
-
-    let tagged = Probe::open(audio_path).ok()?.read().ok()?;
-    let tag = tagged.primary_tag().or_else(|| tagged.first_tag())?;
-    tag.get_string(ItemKey::Lyrics)
-        .filter(|s| !s.trim().is_empty())
-        .map(|s| s.to_string())
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-fn read_local_lrc(audio_path: &str) -> Option<String> {
-    use std::path::Path;
-
-    let audio = Path::new(audio_path);
-
-    let stem_lrc = audio.with_extension("lrc");
-    if let Ok(content) = std::fs::read_to_string(&stem_lrc) {
-        return Some(content);
-    }
-
-    let appended = format!("{audio_path}.lrc");
-    if let Ok(content) = std::fs::read_to_string(&appended) {
-        return Some(content);
-    }
-
-    let parent = audio.parent()?;
-    let file_name = audio.file_name()?.to_string_lossy().to_lowercase();
-    let stem = audio
-        .file_stem()
-        .map(|s| s.to_string_lossy().to_lowercase());
-
-    for entry in std::fs::read_dir(parent).ok()?.flatten() {
-        let path = entry.path();
-        if path
-            .extension()
-            .map(|e| !e.eq_ignore_ascii_case("lrc"))
-            .unwrap_or(true)
-        {
-            continue;
-        }
-        let Some(cand_name) = path.file_name().map(|n| n.to_string_lossy().to_lowercase()) else {
-            continue;
-        };
-        let matches_appended = cand_name == format!("{file_name}.lrc");
-        let matches_stem = stem
-            .as_ref()
-            .map(|s| cand_name == format!("{s}.lrc"))
-            .unwrap_or(false);
-        if (matches_appended || matches_stem)
-            && let Ok(content) = std::fs::read_to_string(&path)
-        {
-            return Some(content);
-        }
-    }
-
-    None
-}
-
-#[cfg(target_arch = "wasm32")]
-async fn fetch_local_lrc(_audio_path: &str) -> Option<Lyrics> {
-    None
-}
-
-async fn fetch_jellyfin_lyrics(item_id: &str, server_url: &str, token: &str) -> Option<Lyrics> {
-    let url = format!(
-        "{}/Items/{}/Lyrics",
-        server_url.trim_end_matches('/'),
-        item_id
-    );
-    let client = reqwest::Client::new();
-    let resp = client
-        .get(&url)
-        .header("X-Emby-Token", token)
-        .timeout(SERVER_LYRICS_TIMEOUT)
-        .send()
-        .await
-        .ok()?;
-
-    if !resp.status().is_success() {
-        return None;
-    }
-
-    let data: JellyfinLyricsResponse = resp.json().await.ok()?;
-    if data.lyrics.is_empty() {
-        return None;
-    }
-
-    let has_timestamps = data.lyrics.iter().any(|l| l.start.is_some());
-    if has_timestamps {
-        let mut lines: Vec<LyricLine> = data
-            .lyrics
-            .iter()
-            .filter(|l| !l.text.trim().is_empty())
-            .filter_map(|l| {
-                l.start.map(|ticks| LyricLine {
-                    // Jellyfin uses 100-nanosecond ticks; divide by 10,000,000 to get seconds
-                    start_time: ticks as f64 / 10_000_000.0,
-                    end_time: None,
-                    text: l.text.clone(),
-                    chunks: Vec::new(),
-                    parent_line_index: None,
-                    background: false,
-                    opposite_turn: false,
-                })
-            })
-            .collect();
-        lines.sort_by(|a, b| {
-            a.start_time
-                .partial_cmp(&b.start_time)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-        if has_usable_line_timing(&lines) {
-            return Some(Lyrics::Synced(lines));
-        }
-    }
-
-    let text: String = data
-        .lyrics
-        .iter()
-        .map(|l| l.text.as_str())
-        .collect::<Vec<_>>()
-        .join("\n");
-    if !text.trim().is_empty() {
-        Some(Lyrics::Plain(text))
-    } else {
-        None
-    }
-}
-
-async fn fetch_subsonic_lyrics(
-    song_id: &str,
-    server_url: &str,
-    username: &str,
-    password: &str,
-    artist: &str,
-    title: &str,
-) -> Option<Lyrics> {
-    let base_url = server_url.trim_end_matches('/');
-    let client = reqwest::Client::new();
-
-    // Subsonic uses MD5(password+salt) for auth
-    let salt = "kopuzlyrics";
-    let token = format!("{:x}", md5::compute(format!("{}{}", password, salt)));
-
-    // Try getLyricsBySongId first (OpenSubsonic extension, synced lyrics)
-    if let Some(lyrics) =
-        subsonic_get_by_id(&client, base_url, username, &token, salt, song_id).await
-    {
-        return Some(lyrics);
-    }
-
-    // Fall back to getLyrics (title+artist, plain text or LRC)
-    subsonic_get_by_title(&client, base_url, username, &token, salt, artist, title).await
-}
-
-async fn subsonic_get_by_id(
-    client: &reqwest::Client,
-    base_url: &str,
-    username: &str,
-    token: &str,
-    salt: &str,
-    song_id: &str,
-) -> Option<Lyrics> {
-    let url = format!("{}/rest/getLyricsBySongId.view", base_url);
-    let params = [
-        ("u", username),
-        ("t", token),
-        ("s", salt),
-        ("v", "1.16.1"),
-        ("c", "kopuz"),
-        ("f", "json"),
-        ("id", song_id),
-    ];
-
-    let resp = client
-        .get(&url)
-        .query(&params)
-        .timeout(SERVER_LYRICS_TIMEOUT)
-        .send()
-        .await
-        .ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-
-    let data: SubsonicSongLyricsEnvelope = resp.json().await.ok()?;
-    if !data.response.status.eq_ignore_ascii_case("ok") {
-        return None;
-    }
-
-    let structured = data
-        .response
-        .lyrics_list?
-        .structured_lyrics
-        .into_iter()
-        .next()?;
-
-    if structured.synced {
-        let mut lines: Vec<LyricLine> = structured
-            .line
-            .iter()
-            .filter(|l| !l.value.trim().is_empty())
-            .filter_map(|l| {
-                l.start.map(|ms| LyricLine {
-                    start_time: ms as f64 / 1000.0,
-                    end_time: None,
-                    text: l.value.clone(),
-                    chunks: Vec::new(),
-                    parent_line_index: None,
-                    background: false,
-                    opposite_turn: false,
-                })
-            })
-            .collect();
-        lines.sort_by(|a, b| {
-            a.start_time
-                .partial_cmp(&b.start_time)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-        if has_usable_line_timing(&lines) {
-            return Some(Lyrics::Synced(lines));
-        }
-    }
-
-    let text: String = structured
-        .line
-        .iter()
-        .map(|l| l.value.as_str())
-        .collect::<Vec<_>>()
-        .join("\n");
-    if !text.trim().is_empty() {
-        Some(Lyrics::Plain(text))
-    } else {
-        None
-    }
-}
-
-async fn subsonic_get_by_title(
-    client: &reqwest::Client,
-    base_url: &str,
-    username: &str,
-    token: &str,
-    salt: &str,
-    artist: &str,
-    title: &str,
-) -> Option<Lyrics> {
-    let url = format!("{}/rest/getLyrics.view", base_url);
-    let params = [
-        ("u", username),
-        ("t", token),
-        ("s", salt),
-        ("v", "1.16.1"),
-        ("c", "kopuz"),
-        ("f", "json"),
-        ("artist", artist),
-        ("title", title),
-    ];
-
-    let resp = client
-        .get(&url)
-        .query(&params)
-        .timeout(SERVER_LYRICS_TIMEOUT)
-        .send()
-        .await
-        .ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-
-    let data: SubsonicPlainLyricsEnvelope = resp.json().await.ok()?;
-    if !data.response.status.eq_ignore_ascii_case("ok") {
-        return None;
-    }
-
-    let value = data.response.lyrics?.value;
-    if value.trim().is_empty() {
-        return None;
-    }
-
-    // getLyrics may return LRC-formatted text
-    let parsed = if lrc_has_usable_timing(&value) {
-        parse_lrc(&value)
-    } else {
-        Vec::new()
-    };
-    if has_usable_line_timing(&parsed) {
-        Some(Lyrics::Synced(parsed))
-    } else {
-        Some(Lyrics::Plain(value))
-    }
-}
-
-async fn fetch_from_musixmatch_enhanced(artist: &str, title: &str) -> Option<Lyrics> {
-    let query = format!("{title} {artist}");
-    let query = query.trim();
-    if query.is_empty() {
-        return None;
-    }
-
-    let client = reqwest::Client::new();
-    let started = Instant::now();
-    let search = musixmatch_get(
-        &client,
-        "track.search",
-        vec![
-            ("q", query.to_string()),
-            ("page_size", "5".to_string()),
-            ("page", "1".to_string()),
-        ],
-        true,
-    )
-    .await?;
-    tracing::info!(
-        target: "kopuz::lyrics",
-        "musixmatch request action=track.search elapsed_ms={}",
-        started.elapsed().as_millis()
-    );
-
-    let status = search.pointer("/message/header/status_code")?.as_i64()?;
-    if status != 200 {
-        tracing::info!(
-            target: "kopuz::lyrics",
-            "musixmatch track.search status={status}"
-        );
-        lyrics_debug!("musixmatch track.search status={status}");
-        return None;
-    }
-
-    let tracks = search.pointer("/message/body/track_list")?.as_array()?;
-    let Some(track_id) = best_musixmatch_track_id(tracks, query) else {
-        tracing::info!(
-            target: "kopuz::lyrics",
-            "musixmatch no_match query={query:?} candidates={}",
-            tracks.len()
-        );
-        lyrics_debug!(
-            "musixmatch no_match candidates={} query={:?}",
-            tracks.len(),
-            query
-        );
-        return None;
-    };
-    lyrics_debug!(
-        "musixmatch selected_track_id={} candidates={}",
-        track_id,
-        tracks.len()
-    );
-    let started = Instant::now();
-    let richsync = musixmatch_get(
-        &client,
-        "track.richsync.get",
-        vec![("track_id", track_id)],
-        true,
-    )
-    .await?;
-    tracing::info!(
-        target: "kopuz::lyrics",
-        "musixmatch request action=track.richsync.get elapsed_ms={}",
-        started.elapsed().as_millis()
-    );
-
-    let status = richsync.pointer("/message/header/status_code")?.as_i64()?;
-    if status != 200 {
-        tracing::info!(
-            target: "kopuz::lyrics",
-            "musixmatch richsync status={status}"
-        );
-        lyrics_debug!("musixmatch richsync status={status}");
-        return None;
-    }
-
-    let body = richsync
-        .pointer("/message/body/richsync/richsync_body")?
-        .as_str()?;
-    let enhanced_lrc = musixmatch_richsync_to_lrc(body)?;
-    if !lrc_has_usable_timing(&enhanced_lrc) {
-        lyrics_debug!("musixmatch richsync has no usable timing");
-        return None;
-    }
-    let parsed = parse_lrc(&enhanced_lrc);
-    if has_usable_line_timing(&parsed) && parsed.iter().any(|line| !line.chunks.is_empty()) {
-        lyrics_debug!(
-            "musixmatch richsync parsed lines={} timed_lines={} timed_parts={}",
-            parsed.len(),
-            timed_line_count(&parsed),
-            timed_part_count(&parsed)
-        );
-        Some(Lyrics::Synced(parsed))
-    } else {
-        lyrics_debug!("musixmatch richsync parsed without word timestamps");
-        None
-    }
-}
-
-async fn fetch_from_paxsenix_youtube(
-    artist: &str,
-    title: &str,
-    duration: u64,
-    track_path: &str,
-) -> Option<Lyrics> {
-    let client = reqwest::Client::new();
-    let video_id = if let Some(video_id) = extract_youtube_video_id(track_path) {
-        video_id
-    } else {
-        let query = format!("{title} {artist}");
-        let query = query.trim();
-        if query.is_empty() {
-            return None;
-        }
-
-        let results = client
-            .get(format!("{PAXSENIX_ROOT_URL}/youtube/search"))
-            .query(&[("q", query)])
-            .timeout(PAXSENIX_TIMEOUT)
-            .send()
-            .await
-            .map_err(|error| {
-                tracing::info!(
-                    target: "kopuz::lyrics",
-                    "paxsenix_youtube search failed={error}"
-                );
-            })
-            .ok()?
-            .json::<Vec<PaxsenixYoutubeSearchResult>>()
-            .await
-            .map_err(|error| {
-                tracing::info!(
-                    target: "kopuz::lyrics",
-                    "paxsenix_youtube search json_failed={error}"
-                );
-            })
-            .ok()?;
-
-        let selected = best_youtube_result(&results, query, duration)?;
-        lyrics_debug!(
-            "paxsenix_youtube selected_video id={} title={:?} artist={:?} candidates={}",
-            selected.video_id,
-            selected.title,
-            selected.author,
-            results.len()
-        );
-        selected.video_id.clone()
-    };
-
-    let lrc = client
-        .get(format!("{PAXSENIX_ROOT_URL}/youtube/lyrics"))
-        .query(&[("id", video_id.as_str())])
-        .timeout(PAXSENIX_YOUTUBE_LYRICS_TIMEOUT)
-        .send()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_youtube lyrics failed={error}"
-            );
-        })
-        .ok()?
-        .text()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_youtube lyrics text_failed={error}"
-            );
-        })
-        .ok()?;
-
-    if lrc.trim().is_empty() || !lrc_has_usable_timing(&lrc) {
-        return None;
-    }
-
-    let lines = parse_lrc(&lrc);
-    if has_usable_line_timing(&lines) {
-        Some(Lyrics::Synced(lines))
-    } else {
-        None
-    }
-}
-
-fn best_youtube_result<'a>(
-    results: &'a [PaxsenixYoutubeSearchResult],
-    query: &str,
-    duration: u64,
-) -> Option<&'a PaxsenixYoutubeSearchResult> {
-    results
-        .iter()
-        .filter_map(|result| {
-            let candidate = format!("{} {}", result.title, result.author);
-            let text_score = lyrics_match_score(&candidate, query);
-            if text_score < 55.0 {
-                return None;
-            }
-
-            let duration_score = match (duration, parse_colon_duration(&result.duration)) {
-                (0, _) | (_, None) => 0.0,
-                (expected, Some(candidate_seconds)) => {
-                    let delta = candidate_seconds.abs_diff(expected);
-                    if delta > 12 {
-                        return None;
-                    }
-                    12.0 - delta as f64
-                }
-            };
-
-            Some((text_score + duration_score, result))
-        })
-        .max_by(|a, b| a.0.partial_cmp(&b.0).unwrap_or(std::cmp::Ordering::Equal))
-        .map(|(_, result)| result)
-}
-
-fn parse_colon_duration(duration: &str) -> Option<u64> {
-    let mut total = 0_u64;
-    let mut parts = duration.split(':').peekable();
-    parts.peek()?;
-    for part in parts {
-        total = total
-            .checked_mul(60)?
-            .checked_add(part.parse::<u64>().ok()?)?;
-    }
-    Some(total)
-}
-
-fn extract_youtube_video_id(track_path: &str) -> Option<String> {
-    track_path
-        .strip_prefix("ytmusic:")
-        .and_then(|rest| rest.split(':').next())
-        .filter(|video_id| !video_id.trim().is_empty())
-        .map(|video_id| video_id.to_string())
-}
-
-async fn fetch_from_paxsenix_apple_music(
-    artist: &str,
-    title: &str,
-    duration: u64,
-) -> Option<Lyrics> {
-    let query = format!("{title} {artist}");
-    let query = query.trim();
-    if query.is_empty() {
-        return None;
-    }
-
-    let client = reqwest::Client::new();
-    let search = client
-        .get(ITUNES_SEARCH_ROOT_URL)
-        .query(&[
-            ("term", query),
-            ("entity", "song"),
-            ("limit", "8"),
-            ("country", "US"),
-        ])
-        .timeout(PAXSENIX_TIMEOUT)
-        .send()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_apple itunes_search failed={error}"
-            );
-        })
-        .ok()?
-        .json::<ItunesSearchResponse>()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_apple itunes_search json_failed={error}"
-            );
-        })
-        .ok()?;
-
-    let Some(song) = best_itunes_song(&search.results, query, duration) else {
-        tracing::info!(
-            target: "kopuz::lyrics",
-            "paxsenix_apple no_match query={query:?} candidates={}",
-            search.results.len()
-        );
-        lyrics_debug!(
-            "paxsenix_apple no_match candidates={} query={:?}",
-            search.results.len(),
-            query
-        );
-        return None;
-    };
-    lyrics_debug!(
-        "paxsenix_apple selected_track id={} title={:?} artist={:?} candidates={}",
-        song.track_id,
-        song.track_name,
-        song.artist_name,
-        search.results.len()
-    );
-
-    let response = client
-        .get(format!("{PAXSENIX_ROOT_URL}/apple-music/lyrics"))
-        .query(&[("id", song.track_id.to_string())])
-        .timeout(PAXSENIX_APPLE_LYRICS_TIMEOUT)
-        .send()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_apple lyrics failed={error}"
-            );
-        })
-        .ok()?
-        .json::<PaxsenixAppleLyricsResponse>()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "paxsenix_apple lyrics json_failed={error}"
-            );
-        })
-        .ok()?;
-
-    let lyrics = paxsenix_apple_to_lyrics(response)?;
-    if let Lyrics::Synced(lines) = &lyrics {
-        lyrics_debug!(
-            "paxsenix_apple parsed kind={} lines={} syllable_lines={} syllable_parts={}",
-            lyrics_kind(Some(&lyrics)),
-            lines.len(),
-            timed_line_count(lines),
-            timed_part_count(lines)
-        );
-    } else {
-        lyrics_debug!("paxsenix_apple parsed kind={}", lyrics_kind(Some(&lyrics)));
-    }
-    Some(lyrics)
-}
-
-fn best_itunes_song<'a>(
-    songs: &'a [ItunesSong],
-    query: &str,
-    duration: u64,
-) -> Option<&'a ItunesSong> {
-    songs
-        .iter()
-        .filter_map(|song| {
-            let candidate = format!("{} {}", song.track_name, song.artist_name);
-            let text_score = lyrics_match_score(&candidate, query);
-            if text_score < 55.0 {
-                return None;
-            }
-
-            let duration_score = match (duration, song.track_time_millis) {
-                (0, _) | (_, None) => 0.0,
-                (expected, Some(candidate_ms)) => {
-                    let candidate_seconds = candidate_ms / 1000;
-                    let delta = candidate_seconds.abs_diff(expected);
-                    if delta > 12 {
-                        return None;
-                    }
-                    12.0 - delta as f64
-                }
-            };
-
-            Some((text_score + duration_score, song))
-        })
-        .max_by(|a, b| a.0.partial_cmp(&b.0).unwrap_or(std::cmp::Ordering::Equal))
-        .map(|(_, song)| song)
-}
-
-fn paxsenix_apple_to_lyrics(response: PaxsenixAppleLyricsResponse) -> Option<Lyrics> {
-    let lines = paxsenix_apple_to_lines(response.content);
-    if let Some(lines) = lines
-        && has_usable_line_timing(&lines)
-    {
-        return Some(Lyrics::Synced(lines));
-    }
-
-    if let Some(lrc) = response.lrc
-        && !lrc.trim().is_empty()
-    {
-        let parsed = if lrc_has_usable_timing(&lrc) {
-            parse_lrc(&lrc)
-        } else {
-            Vec::new()
-        };
-        if has_usable_line_timing(&parsed) {
-            return Some(Lyrics::Synced(parsed));
-        }
-    }
-
-    response
-        .plain
-        .filter(|plain| !plain.trim().is_empty())
-        .map(Lyrics::Plain)
-}
-
-fn paxsenix_apple_to_lines(rows: Vec<PaxsenixAppleLyricLine>) -> Option<Vec<LyricLine>> {
-    if !paxsenix_apple_has_timing(&rows) {
-        lyrics_debug!("paxsenix_apple content has no usable timing");
-        return None;
-    }
-
-    let mut lines = Vec::new();
-
-    for row in rows {
-        let row_start_time = row.timestamp as f64 / 1000.0;
-        let row_end_time = row.endtime.map(|endtime| endtime as f64 / 1000.0);
-        let opposite_turn = row.opposite_turn;
-        let row_has_background_text = !row.background_text.is_empty();
-        let main_line_is_background = row.background && !row_has_background_text;
-        let mut parent_line_index = None;
-
-        if let Some(line) = paxsenix_apple_parts_to_line(
-            row.text,
-            row_start_time,
-            row_end_time,
-            None,
-            main_line_is_background,
-            opposite_turn,
-        ) {
-            lines.push(line);
-            if !main_line_is_background {
-                parent_line_index = lines.len().checked_sub(1);
-            }
-        }
-
-        let background_start_time = row
-            .background_text
-            .iter()
-            .find_map(|part| part.timestamp)
-            .map(|timestamp| timestamp as f64 / 1000.0)
-            .unwrap_or(row_start_time);
-
-        if let Some(line) = paxsenix_apple_parts_to_line(
-            row.background_text,
-            background_start_time,
-            row_end_time,
-            parent_line_index,
-            true,
-            opposite_turn,
-        ) {
-            lines.push(line);
-        }
-    }
-
-    (!lines.is_empty()).then_some(lines)
-}
-
-fn paxsenix_apple_has_timing(rows: &[PaxsenixAppleLyricLine]) -> bool {
-    rows.iter().any(|row| {
-        row.timestamp > 0
-            || row.endtime.is_some_and(|endtime| endtime > 0)
-            || row
-                .text
-                .iter()
-                .any(|part| part.timestamp.is_some_and(|timestamp| timestamp > 0))
-            || row
-                .background_text
-                .iter()
-                .any(|part| part.timestamp.is_some_and(|timestamp| timestamp > 0))
-    })
-}
-
-fn paxsenix_apple_parts_to_line(
-    parts: Vec<PaxsenixAppleLyricPart>,
-    start_time: f64,
-    end_time: Option<f64>,
-    parent_line_index: Option<usize>,
-    background: bool,
-    opposite_turn: bool,
-) -> Option<LyricLine> {
-    let mut text = String::new();
-    let mut chunks = Vec::new();
-    let mut previous_part_continues = false;
-
-    for part in parts {
-        if part.text.trim().is_empty() {
-            continue;
-        }
-
-        let prefix = if should_insert_apple_space(&text, previous_part_continues, &part.text) {
-            " "
-        } else {
-            ""
-        };
-        let display_text = format!("{prefix}{}", part.text);
-
-        text.push_str(&display_text);
-        if let Some(timestamp) = part.timestamp {
-            chunks.push(LyricChunk {
-                start_time: timestamp as f64 / 1000.0,
-                text: display_text,
-            });
-        }
-        previous_part_continues = part.part;
-    }
-
-    let text = text.trim().to_string();
-    (!text.is_empty()).then_some(LyricLine {
-        start_time,
-        end_time,
-        text,
-        chunks,
-        parent_line_index,
-        background,
-        opposite_turn,
-    })
-}
-
-fn should_insert_apple_space(
-    current_text: &str,
-    previous_part_continues: bool,
-    next_text: &str,
-) -> bool {
-    if current_text.is_empty() || previous_part_continues {
-        return false;
-    }
-
-    let Some(first_char) = next_text.chars().next() else {
-        return false;
-    };
-
-    !matches!(
-        first_char,
-        ',' | '.' | '?' | '!' | ':' | ';' | ')' | ']' | '}' | '\'' | '’'
-    )
-}
-
-async fn musixmatch_get(
-    client: &reqwest::Client,
-    action: &str,
-    mut query: Vec<(&str, String)>,
-    needs_token: bool,
-) -> Option<serde_json::Value> {
-    if needs_token {
-        let token = musixmatch_token(client).await?;
-        query.push(("usertoken", token));
-    }
-
-    query.push(("app_id", "web-desktop-app-v1.0".to_string()));
-    query.push(("t", now_ms().to_string()));
-
-    client
-        .get(format!("{MUSIXMATCH_ROOT_URL}{action}"))
-        .query(&query)
-        .timeout(MUSIXMATCH_TIMEOUT)
-        .send()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "musixmatch request action={action} failed={error}"
-            );
-        })
-        .ok()?
-        .json::<serde_json::Value>()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "musixmatch request action={action} json_failed={error}"
-            );
-        })
-        .ok()
-}
-
-async fn musixmatch_token(client: &reqwest::Client) -> Option<String> {
-    let now = now_ms();
-    if let Some(token) = musixmatch_token_cache().lock().ok().and_then(|cache| {
-        cache
-            .as_ref()
-            .filter(|token| token.expires_at_ms > now)
-            .map(|token| token.value.clone())
-    }) {
-        return Some(token);
-    }
-
-    let response = client
-        .get(format!("{MUSIXMATCH_ROOT_URL}token.get"))
-        .query(&[
-            ("user_language", "en".to_string()),
-            ("app_id", "web-desktop-app-v1.0".to_string()),
-            ("t", now.to_string()),
-        ])
-        .timeout(MUSIXMATCH_TIMEOUT)
-        .send()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "musixmatch token request failed={error}"
-            );
-        })
-        .ok()?
-        .json::<serde_json::Value>()
-        .await
-        .map_err(|error| {
-            tracing::info!(
-                target: "kopuz::lyrics",
-                "musixmatch token json_failed={error}"
-            );
-        })
-        .ok()?;
-
-    let status = response.pointer("/message/header/status_code")?.as_i64()?;
-    if status != 200 {
-        tracing::info!(
-            target: "kopuz::lyrics",
-            "musixmatch token status={status}"
-        );
-        return None;
-    }
-
-    let token = response
-        .pointer("/message/body/user_token")?
-        .as_str()?
-        .to_string();
-
-    if let Ok(mut cache) = musixmatch_token_cache().lock() {
-        *cache = Some(MusixmatchToken {
-            value: token.clone(),
-            expires_at_ms: now + 10 * 60 * 1000,
-        });
-    }
-
-    Some(token)
-}
-
-fn musixmatch_token_cache() -> &'static Mutex<Option<MusixmatchToken>> {
-    MUSIXMATCH_TOKEN.get_or_init(|| Mutex::new(None))
-}
-
-fn best_musixmatch_track_id(tracks: &[serde_json::Value], query: &str) -> Option<String> {
-    tracks
-        .iter()
-        .filter_map(|entry| {
-            let track = entry.get("track")?;
-            let name = track.get("track_name")?.as_str().unwrap_or_default();
-            let artist = track.get("artist_name")?.as_str().unwrap_or_default();
-            let candidate = format!("{name} {artist}");
-            let score = lyrics_match_score(&candidate, query);
-            let id = track.get("track_id")?.as_i64()?.to_string();
-            Some((score, id))
-        })
-        .max_by(|a, b| a.0.partial_cmp(&b.0).unwrap_or(std::cmp::Ordering::Equal))
-        .and_then(|(score, id)| (score >= 65.0).then_some(id))
-}
-
 fn lyrics_match_score(candidate: &str, query: &str) -> f64 {
     let candidate_tokens = normalized_lyric_match_tokens(candidate);
     let query_tokens = normalized_lyric_match_tokens(query);
@@ -1868,59 +828,6 @@ fn normalized_lyric_match_tokens(value: &str) -> HashSet<String> {
         .filter(|token| !token.is_empty())
         .map(ToString::to_string)
         .collect()
-}
-
-fn musixmatch_richsync_to_lrc(body: &str) -> Option<String> {
-    let rows = serde_json::from_str::<Vec<serde_json::Value>>(body).ok()?;
-    let mut output = String::new();
-
-    for row in rows {
-        let line_start = json_number(&row, "ts")?;
-        output.push('[');
-        output.push_str(&format_lrc_time(line_start));
-        output.push(']');
-
-        if let Some(words) = row.get("l").and_then(|value| value.as_array()) {
-            for word in words {
-                let offset = json_number(word, "o").unwrap_or(0.0);
-                let content = word.get("c").and_then(|value| value.as_str()).unwrap_or("");
-                if content.trim().is_empty() {
-                    continue;
-                }
-                output.push('<');
-                output.push_str(&format_lrc_time(line_start + offset));
-                output.push('>');
-                output.push_str(content);
-                output.push(' ');
-            }
-        }
-
-        output.push('\n');
-    }
-
-    (!output.trim().is_empty()).then_some(output)
-}
-
-fn json_number(value: &serde_json::Value, key: &str) -> Option<f64> {
-    value
-        .get(key)
-        .and_then(|number| number.as_f64().or_else(|| number.as_str()?.parse().ok()))
-}
-
-fn format_lrc_time(time_in_seconds: f64) -> String {
-    let time = time_in_seconds.max(0.0);
-    let total_centiseconds = (time * 100.0).round() as u64;
-    let minutes = total_centiseconds / 6000;
-    let seconds = (total_centiseconds / 100) % 60;
-    let centiseconds = total_centiseconds % 100;
-    format!("{minutes:02}:{seconds:02}.{centiseconds:02}")
-}
-
-fn now_ms() -> u128 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_millis())
-        .unwrap_or_default()
 }
 
 async fn fetch_from_lrclib(
@@ -2028,202 +935,13 @@ fn extract_from_lrclib_response(data: &LrcLibResponse) -> Option<Lyrics> {
     None
 }
 
-fn append_translation(target: &mut String, text: &str) {
-    let text = text.trim();
-
-    if text.is_empty() {
-        return;
-    }
-
-    if !target.is_empty() {
-        target.push('\n');
-    }
-
-    if text.starts_with('(') && text.ends_with(')') {
-        target.push_str(text);
-    } else {
-        target.push('(');
-        target.push_str(text);
-        target.push(')');
-    }
-}
-
-fn append_line_translation(target: &mut LyricLine, text: &str) {
-    append_translation(&mut target.text, text);
-}
-
-fn parse_lrc(lrc_text: &str) -> Vec<LyricLine> {
-    let mut lines: Vec<LyricLine> = Vec::new();
-
-    for raw_line in lrc_text.lines() {
-        let (line_timestamps, content) = extract_line_timestamps(raw_line);
-        let (text, words) = parse_enhanced_words(content);
-
-        if line_timestamps.is_empty() {
-            let is_metadata_tag = raw_line.trim().starts_with('[')
-                && raw_line.trim().ends_with(']')
-                && raw_line.contains(':');
-            if is_metadata_tag {
-                continue;
-            }
-
-            if !words.is_empty() {
-                lines.push(LyricLine {
-                    start_time: words[0].start_time,
-                    end_time: None,
-                    text,
-                    chunks: words,
-                    parent_line_index: None,
-                    background: false,
-                    opposite_turn: false,
-                });
-            } else if !text.is_empty()
-                && let Some(last) = lines.last_mut()
-            {
-                append_line_translation(last, &text);
-            }
-            continue;
-        }
-
-        if text.is_empty() && words.is_empty() {
-            continue;
-        }
-
-        for start_time in line_timestamps {
-            lines.push(LyricLine {
-                start_time,
-                end_time: None,
-                text: text.clone(),
-                chunks: words.clone(),
-                parent_line_index: None,
-                background: false,
-                opposite_turn: false,
-            });
-        }
-    }
-
-    lines.sort_by(|a, b| {
-        a.start_time
-            .partial_cmp(&b.start_time)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-
-    let mut merged: Vec<LyricLine> = Vec::new();
-
-    for mut line in lines {
-        if let Some(last) = merged.last_mut()
-            && last.start_time == line.start_time
-        {
-            if last.chunks.is_empty() && !line.chunks.is_empty() {
-                last.chunks = std::mem::take(&mut line.chunks);
-            }
-            if last.end_time.is_none() {
-                last.end_time = line.end_time;
-            }
-            append_line_translation(last, &line.text);
-            continue;
-        }
-
-        merged.push(line);
-    }
-
-    merged
-}
-
-fn extract_line_timestamps(line: &str) -> (Vec<f64>, &str) {
-    let mut timestamps = Vec::new();
-    let mut rest = line.trim_start();
-
-    while let Some(after_open) = rest.strip_prefix('[') {
-        let Some(close_idx) = after_open.find(']') else {
-            break;
-        };
-
-        let tag = &after_open[..close_idx];
-        let after_tag = &after_open[close_idx + 1..];
-        if let Some(time) = parse_lrc_time(tag) {
-            timestamps.push(time);
-            rest = after_tag;
-        } else if tag.chars().any(|c| c.is_ascii_alphabetic()) && tag.contains(':') {
-            rest = after_tag;
-        } else {
-            break;
-        }
-    }
-
-    (timestamps, rest)
-}
-
-fn parse_enhanced_words(content: &str) -> (String, Vec<LyricChunk>) {
-    let mut words = Vec::new();
-    let mut text = String::new();
-    let mut rest = content;
-    let mut pending_time: Option<f64> = None;
-
-    while let Some(open_idx) = rest.find('<') {
-        let before = &rest[..open_idx];
-        text.push_str(before);
-        if let Some(start_time) = pending_time.take()
-            && !before.is_empty()
-        {
-            words.push(LyricChunk {
-                start_time,
-                text: before.to_string(),
-            });
-        }
-
-        let after_open = &rest[open_idx + 1..];
-        let Some(close_idx) = after_open.find('>') else {
-            text.push_str(&rest[open_idx..]);
-            rest = "";
-            break;
-        };
-
-        let tag = &after_open[..close_idx];
-        if let Some(time) = parse_lrc_time(tag) {
-            pending_time = Some(time);
-            rest = &after_open[close_idx + 1..];
-        } else {
-            text.push('<');
-            text.push_str(tag);
-            text.push('>');
-            rest = &after_open[close_idx + 1..];
-        }
-    }
-
-    text.push_str(rest);
-    if let Some(start_time) = pending_time
-        && !rest.is_empty()
-    {
-        words.push(LyricChunk {
-            start_time,
-            text: rest.to_string(),
-        });
-    }
-
-    (text.trim().to_string(), words)
-}
-
-fn parse_lrc_time(time_str: &str) -> Option<f64> {
-    let parts: Vec<&str> = time_str.split([':', '.']).collect();
-    if parts.len() >= 2 {
-        let min = parts[0].parse::<f64>().ok()?;
-        let sec = parts[1].parse::<f64>().ok()?;
-        let mut total = min * 60.0 + sec;
-        if parts.len() == 3 {
-            let ms_str = parts[2];
-            let ms = ms_str.parse::<f64>().ok()?;
-            let divisor = 10_f64.powi(ms_str.len() as i32);
-            total += ms / divisor;
-        }
-        Some(total)
-    } else {
-        None
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use super::musixmatch::musixmatch_richsync_to_lrc;
+    use super::paxsenix::{
+        best_itunes_song, best_youtube_result, extract_youtube_video_id, parse_colon_duration,
+        paxsenix_apple_to_lyrics,
+    };
     use super::*;
 
     #[test]
